fix: macOS platform compatibility - float-to-double conversions and includes
- Add explicit static_cast<double>() for float-to-qreal/double conversions
  throughout codebase to fix Apple Clang -Wdouble-promotion strict warnings
- Add missing Qt headers: QStyleHints, QPointer, QGuiApplication
- Add [[maybe_unused]] attributes for platform-specific joystick code that's
  only used inside Linux ifdef guards
- Remove unused lambda captures (scenarioLogLine, unused in some contexts)
- Suppress -Wc++20-extensions flag (structured bindings only needed for storage,
  not capture, and are caught by C++17 guard in code)

These fixes fixes compilation on macOS while maintaining code quality standards.
Platform-specific guards remain for Linux-only features (backtrace, joystick).. have to figure out how to add joystcks for macos via their embedded APIs as it helps mapping better than the /dev/joy. readouts.

Fixes compilation errors on macOS without changing build behavior on Linux.

+++ src/vehicle/import/kit_bundler.cpp
@@ -39,8 +39,8 @@ QString render_kit_readme(const VehicleSpec &spec) {
   s << "| Field | Value |\n|---|---|\n";
   s << "| Name | `" << spec.name << "` |\n";
   s << "| Size class | `" << size_class_name(spec.size_class) << "` |\n";
-  s << "| Mass | " << QString::number(spec.mass, 'f', 0) << " kg |\n";
-  s << "| Suspension damping | " << QString::number(spec.susp_damping, 'f', 2) << " |\n";
+  s << "| Mass | " << QString::number(static_cast<double>(spec.mass), 'f', 0) << " kg |\n";
+  s << "| Suspension damping | " << QString::number(static_cast<double>(spec.susp_damping), 'f', 2) << " |\n";
   s << "| Torque-curve preset | `" << spec.torque_curve_tag << "` |\n";
   s << "| Chassis bbox (cm) | "
     << QString("X[%1..%2] Y[%3..%4] Z[%5..%6]")
@@ -58,11 +58,11 @@ QString render_kit_readme(const VehicleSpec &spec) {
   for (size_t i = 0; i < 4; ++i) {
     const WheelSpec &w = spec.wheels[i];
     s << "| " << names[i] << " | "
-      << QString::number(w.x, 'f', 1) << " | "
-      << QString::number(w.y, 'f', 1) << " | "
-      << QString::number(w.z, 'f', 1) << " | "
-      << QString::number(w.radius, 'f', 1) << " | "
-      << QString::number(w.width, 'f', 1) << " |\n";
+      << QString::number(static_cast<double>(w.x), 'f', 1) << " | "
+      << QString::number(static_cast<double>(w.y), 'f', 1) << " | "
+      << QString::number(static_cast<double>(w.z), 'f', 1) << " | "
+      << QString::number(static_cast<double>(w.radius), 'f', 1) << " | "
+      << QString::number(static_cast<double>(w.width), 'f', 1) << " |\n";
   }
   return out;
 }
